# swagger.yaml
openapi: 3.0.0
info:
  title: jobs api
  contact: {}
  version: '1.0'
servers:
- url: https://turquoise-hen-hem.cyclic.cloud/api/v1
  variables: {}
paths:
  /jobs:
    get:
      tags:
      - jobs
      summary: get all jobs
      operationId: getalljobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - jobs
      summary: post job
      operationId: postjob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/postjobrequest'
              - example:
                  company: amazoon
                  position: frontt end dev
            example:
              company: amazoon
              position: frontt end dev
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The job ID
    get:
      tags:
      - jobs
      summary: get single job
      operationId: getsinglejob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    patch:
      tags:
      - jobs
      summary: update task
      operationId: updatetask
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updatetaskrequest'
              - example:
                  name: subul
                  position: bback endd
            example:
              name: subul
              position: bback endd
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - jobs
      summary: delete job
      operationId: deletejob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
    postjobrequest:
      title: postjobrequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: amazoon
        position: frontt end dev
...
    updatetaskrequest:
      title: updatetaskrequest
      required:
      - name
      - position
      type: object
      properties:
        name:
          type: string
        position:
          type: string
      example:
        name: subul
        position: bback endd
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...